+++ pyramid_metrics/tween.py
@@ -16,26 +16,30 @@ def includeme(config):
 def performance_tween_factory(handler, registry):
 
     def performance_tween(request):
+        http_method = request.method.lower()
         try:
             request.metrics.mark_start('request')
             response = handler(request)
         except:
-            request.metrics.mark_stop('request', suffix='.exc')
+            request.metrics.mark_stop('request', suffix=(http_method, 'exc'))
             raise
 
         else:
             if 200 <= response.status_code < 300:
-                suffix = '.2xx'
+                status_code = '2xx'
             elif 300 <= response.status_code < 400:
-                suffix = '.3xx'
+                status_code = '3xx'
             elif 400 <= response.status_code < 500:
-                suffix = '.4xx'
+                status_code = '4xx'
             elif 500 <= response.status_code < 600:
-                suffix = '.5xx'
+                status_code = '5xx'
             else:
-                suffix = '.xxx'
-            request.metrics.mark_stop('request', suffix=suffix)
-            # request.metrics.mark_stop('request', suffix='.total')
+                status_code = 'xxx'
+
+            request.metrics.mark_stop(
+                'request', suffix=(http_method, status_code))
+            # request.metrics.mark_stop(
+            #     'request', suffix=(http_method, 'total'))
 
             return response