+++ pyramid_metrics/utility.py
@@ -61,25 +61,34 @@ class MetricsUtility(object):
     def _route_key(self, stat):
         return 'route.%s.%s' % (self.route_name, stat)
 
+    def _key(self, stat):
+        if isinstance(stat, (list, tuple)):
+            return '.'.join([str(member) for member in stat])
+        else:
+            return str(stat)
+
     def incr(self, stat, count=1, per_route=False):
         """ Push a COUNTER metric
 
-        >>> request.metrics.incr('mysql_reconnection')
+        >>> request.metrics.incr('mysql.reconnection')
+        >>> request.metrics.incr(['mysql', 'reconnection'])
         """
-        self._statsd.incr(stat, count=count)
+        self._statsd.incr(self._key(stat), count=count)
         if per_route:
-            self._statsd.incr(self._route_key(stat), count=count)
+            self._statsd.incr(self._route_key(self._key(stat)),
+                              count=count)
 
     def timing(self, stat, dt, per_route=False):
         """ Push a TIMER metric (in milliseconds)
 
         >>> start_time = time()
         >>> ...
-        >>> request.metrics.timing('sql_select_user', time() - start_time)
+        >>> request.metrics.timing('mysql.select', time() - start_time)
+        >>> request.metrics.timing(['mysql', 'select'], time() - start_time)
         """
-        self._statsd.timing(stat, dt)
+        self._statsd.timing(self._key(stat), dt)
         if per_route:
-            self._statsd.timing(self._route_key(stat), dt)
+            self._statsd.timing(self._route_key(self._key(stat)), dt)
 
     def mark_start(self, marker_name):
         """ Place a start time marker.
@@ -108,7 +117,13 @@ class MetricsUtility(object):
         """
         stop_marker = Marker(marker_name)
         start_marker = self.active_markers.get(marker_name)
+
         if start_marker:
-            stat_key = '%s%s%s' % (prefix, marker_name, suffix)
+            stat = [marker_name]
+            if prefix:
+                stat.insert(0, self._key(prefix))
+            if suffix:
+                stat.append(self._key(suffix))
             dt = stop_marker.get_delta_from(start_marker)
-            self.timing(stat_key, dt, per_route=per_route)
+
+            self.timing(stat, dt=dt, per_route=per_route)